add component and fix
