+++ src/components/Commons/Card/Card.modules.js
@@ -1,73 +1,8 @@
-import styled from "styled-components";
-
-export const BlockPeopleTask = styled.div`
-  max-width: 880px;
-  margin: 0 auto;
-  padding: 0;
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  padding: 0;
-  @media (max-width: 768px) {
-    max-width: 580px;
-  }
-  @media (max-width: 320px) {
-    max-width: 280px;
-  }
-`;
+import styled from 'styled-components';
 
 export const BlockPeopleTask_item = styled.div`
-  max-width: 280px;
-  margin-bottom: 20px;
-  position: relative;
-  list-style: none;
-`;
-
-export const BlockPeopleTask_item_img = styled.div`
-  border-radius: 6px 6px 0px 0px;
-  height: 194px;
-  width: 100%;
-`;
-export const BlockPeopleTask_item_inform = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding-bottom: 14px;
-`;
-
-export const BlockPeopleTask_item_inform_pad = styled.p`
-  /* height: 18px; */
-  padding: 3px 10px;
-  background-color: #5679d7;
-  border-radius: 3px;
-
-  font-style: normal;
-  font-weight: 500;
-  font-size: 10px;
-
-  text-align: center;
-
-  text-transform: uppercase;
-
-  color: #ffffff;
-`;
-export const BlockPeopleTask_item_inform_title = styled.h2`
-  padding-top: 14px;
-  margin-bottom: 4px;
-  text-align: left;
-  font-family: Montserrat;
-  font-style: normal;
-  font-weight: bold;
-  font-size: 12px;
-  line-height: 15px;
-  letter-spacing: 0.04em;
-  text-transform: uppercase;
-
-  color: #000000;
-`;
-
-export const BlockPeopleTask_item_text = styled.div`
-  background: #ffbc33;
-  padding: 0 10px;
-  border-radius: 0px 0px 6px 6px;
+	max-width: 280px;
+	margin-bottom: 20px;
+	position: relative;
+	list-style: none;
 `;
