+++ src/components/Commons/CardBody/CardBody.module.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const BlockPeopleTask_item_img = styled.div`
+	border-radius: 6px 6px 0px 0px;
+	height: 194px;
+	width: 100%;
+`;
