+++ src/components/Commons/CardFooter/CardFooter.module.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const BlockPeopleTask_item_inform = styled.div`
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	padding-bottom: 14px;
+`;
+
+export const BlockPeopleTask_item_text = styled.div`
+	background: #ffbc33;
+	padding: 0 10px;
+	border-radius: 0px 0px 6px 6px;
+`;
