+++ src/components/Commons/CardTitle/CardTitle.js
@@ -1,0 +1,6 @@
+//styles
+import { BlockPeopleTask_item_inform_title } from './CardTitle.module';
+
+export default function CardTitle({ item }) {
+	return <BlockPeopleTask_item_inform_title>{item.title}</BlockPeopleTask_item_inform_title>;
+}
