+++ src/components/Commons/CardTitle/CardTitle.module.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const BlockPeopleTask_item_inform_title = styled.h2`
+	padding-top: 14px;
+	margin-bottom: 4px;
+	text-align: left;
+	font-family: 'Montserrat';
+	font-style: normal;
+	font-weight: bold;
+	font-size: 12px;
+	line-height: 15px;
+	letter-spacing: 0.04em;
+	text-transform: uppercase;
+	color: #000000;
+`;
