+++ src/components/Commons/DaysList/DaysList.styles.js
@@ -1,73 +1,57 @@
-import styled from "styled-components";
+import styled from 'styled-components';
 
-export const BlockPeopleTask = styled.ul`
-  max-width: 1180px;
-  margin: 0 auto;
-  padding: 0;
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-
-  @media (max-width: 768px) {
-    max-width: 580px;
-  }
-  @media (max-width: 320px) {
-    max-width: 280px;
-  }
+export const BlockAddSwitch = styled.div`
+	cursor: pointer;
+	width: 30px;
+	height: 30px;
+	background: #ffffff;
+	border-radius: 50%;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	font-family: 'Montserrat';
+	font-style: normal;
+	font-weight: bold;
+	font-size: 12px;
+	line-height: 15px;
+	display: flex;
+	align-items: center;
+	text-align: center;
+	letter-spacing: 0.04em;
+	color: #8ec63f;
 `;
 
-export const BlockPeopleTask_item = styled.li`
-  max-width: 280px;
-  margin-bottom: 20px;
-  position: relative;
-  list-style: none;
+export const BlockCheckbox = styled.ul`
+	padding: 0;
+	margin: 0;
+	bottom: 70px;
+	right: -20px;
+	position: absolute;
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
+	width: 114px;
+	height: auto;
+	background: #3e7adc;
+	border-radius: 6px;
+	padding-top: 15px;
 `;
 
-export const BlockPeopleTask_img = styled.div`
-  border-radius: 6px 6px 0px 0px;
-  height: 194px;
-  width: 100%;
+export const BlockLabel = styled.li`
+	list-style: none;
+	margin-bottom: 7px;
+	font-family: 'Montserrat';
+	font-style: normal;
+	font-weight: bold;
+	font-size: 14px;
+	line-height: 17px;
+	display: flex;
+	align-items: center;
+	letter-spacing: 0.02em;
+	color: #ffffff;
 `;
 
-export const BlockPeopleTask_item_inform = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding-bottom: 14px;
-`;
-
-export const BlockPeopleTask_item_inform_pad = styled.p`
-  padding: 0;
-  margin: 0;
-  padding: 3px 10px;
-  background-color: #5679d7;
-  border-radius: 3px;
-  font-style: normal;
-  font-weight: 500;
-  font-size: 10px;
-  text-align: center;
-  text-transform: uppercase;
-  color: #ffffff;
-`;
-export const BlockPeopleTask_item_inform_title = styled.h2`
-  padding: 0;
-  margin: 0;
-  padding-top: 14px;
-  margin-bottom: 4px;
-  text-align: left;
-  font-family: Montserrat;
-  font-style: normal;
-  font-weight: bold;
-  font-size: 12px;
-  line-height: 15px;
-  letter-spacing: 0.04em;
-  text-transform: uppercase;
-
-  color: #000000;
-`;
-
-export const BlockPeopleTask_item_text = styled.div`
-  background: #ffbc33;
-  padding: 0 10px;
-  border-radius: 0px 0px 6px 6px;
+export const BlockInput = styled.input`
+	margin-right: 10px;
 `;
