--- src/components/Commons/DaysList/style.module.css
@@ -1,75 +1,0 @@
-.blockPeopleTask {
-  max-width: 1180px;
-  margin: 0 auto;
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-}
-
-.blockPeopleTask_item {
-  max-width: 280px;
-  margin-bottom: 20px;
-  position: relative;
-}
-
-.blockPeopleTask_item_img {
-  border-radius: 6px 6px 0px 0px;
-  height: 194px;
-  width: 100%;
-}
-
-.blockPeopleTask_item_inform {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding-bottom: 14px;
-}
-
-.blockPeopleTask_item_inform > p {
-  /* height: 18px; */
-  padding: 3px 10px;
-  background-color: #5679d7;
-  border-radius: 3px;
-
-  font-style: normal;
-  font-weight: 500;
-  font-size: 10px;
-
-  text-align: center;
-
-  text-transform: uppercase;
-
-  color: #ffffff;
-}
-
-.blockPeopleTask_item_text > h2 {
-  padding-top: 14px;
-  margin-bottom: 4px;
-  text-align: left;
-  font-family: Montserrat;
-  font-style: normal;
-  font-weight: bold;
-  font-size: 12px;
-  line-height: 15px;
-  letter-spacing: 0.04em;
-  text-transform: uppercase;
-
-  color: #000000;
-}
-
-.blockPeopleTask_item_text {
-  background: #ffbc33;
-  padding: 0 10px;
-  border-radius: 0px 0px 6px 6px;
-}
-
-@media (max-width: 768px) {
-  .blockPeopleTask {
-    max-width: 580px;
-  }
-}
-@media (max-width: 320px) {
-  .blockPeopleTask {
-    max-width: 280px;
-  }
-}
