+++ src/components/Commons/PointAmount/PointAmount.js
@@ -1,0 +1,6 @@
+//styles
+import { BlockPeopleTask_item_inform_pad } from './PointAmount.module';
+
+export default function PointAmount({ item }) {
+	return <BlockPeopleTask_item_inform_pad>{item.ball}</BlockPeopleTask_item_inform_pad>;
+}
