+++ src/components/Commons/PointAmount/PointAmount.module.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const BlockPeopleTask_item_inform_pad = styled.p`
+	padding: 3px 10px;
+	background-color: #5679d7;
+	border-radius: 3px;
+	font-style: normal;
+	font-weight: 500;
+	font-size: 10px;
+	text-align: center;
+	text-transform: uppercase;
+	color: #ffffff;
+`;
