+++ src/components/Commons/TaskToggle/SvgIconSwitch.js
@@ -1,0 +1,24 @@
+export const Unchecked = () => {
+	return (
+		<svg width="4" height="22" viewBox="0 0 4 13" fill="none" xmlns="http://www.w3.org/2000/svg">
+			<path
+				d="M0.832 0.799999H3.808L3.312 8.112H1.328L0.832 0.799999ZM2.32 12.128C1.88267 12.128 1.51467 11.9893 1.216 11.712C0.928 11.424 0.784 11.0773 0.784 10.672C0.784 10.2667 0.928 9.93067 1.216 9.664C1.504 9.38667 1.872 9.248 2.32 9.248C2.768 9.248 3.136 9.38667 3.424 9.664C3.712 9.93067 3.856 10.2667 3.856 10.672C3.856 11.0773 3.70667 11.424 3.408 11.712C3.12 11.9893 2.75733 12.128 2.32 12.128Z"
+				fill="white"
+			/>
+		</svg>
+	);
+};
+
+export const Checked = () => {
+	return (
+		<svg width="14" height="10" viewBox="0 0 12 10" fill="none" xmlns="http://www.w3.org/2000/svg">
+			<path
+				d="M1 4.85185L4.42857 9L11 1"
+				stroke="white"
+				strokeWidth="2"
+				strokeLinecap="round"
+				strokeLinejoin="round"
+			/>
+		</svg>
+	);
+};
